Zadania 1-3 wykonane
