+++ src/main/java/edu/kis/vh/nursery/HanoiRhymer.java
@@ -1,13 +1,13 @@
 package edu.kis.vh.nursery;
 
-public class HanoiRhymer extends defaultCountingOutRhymer {
+public class HanoiRhymer extends DefaultCountingOutRhymer {
 
 int totalRejected = 0;
 
 	public int reportRejected() {
 		return totalRejected;
 	}
-
+	@Overide
 	public void countIn(int in) {
 	if (!callCheck() && in > peekaboo())
 			totalRejected++;
